refactor: use llm handler
