+++ src/libs/llm/base.ts
@@ -1,7 +1,10 @@
+import type { NeuriContext } from 'neuri'
+import type { ChatCompletion, Message } from 'neuri/openai'
 import type { LLMConfig, LLMResponse } from './types'
 
 import { useLogg } from '@guiiai/logg'
 
+import { openaiConfig } from '../../composables/config'
 import { toRetriable } from '../../utils/helper'
 
 export abstract class BaseLLMHandler {
@@ -10,13 +13,13 @@ export abstract class BaseLLMHandler {
   constructor(protected config: LLMConfig) {}
 
   protected async handleCompletion(
-    context: any,
+    context: NeuriContext,
     route: string,
-    messages: any[],
+    messages: Message[],
   ): Promise<LLMResponse> {
     const completion = await context.reroute(route, messages, {
-      model: this.config.model ?? 'openai/gpt-4-mini',
-    })
+      model: this.config.model ?? openaiConfig.model,
+    }) as ChatCompletion | ChatCompletion & { error: { message: string } }
 
     if (!completion || 'error' in completion) {
       this.logger.withFields(context).error('Completion failed')
@@ -32,8 +35,8 @@ export abstract class BaseLLMHandler {
     }
   }
 
-  protected createRetryHandler<T>(handler: (context: any) => Promise<T>) {
-    return toRetriable<any, T>(
+  protected createRetryHandler<T>(handler: (context: NeuriContext) => Promise<T>) {
+    return toRetriable<NeuriContext, T>(
       this.config.retryLimit ?? 3,
       this.config.delayInterval ?? 1000,
       handler,
